refactor: hide routes components directories from router
