+++ app/search/_components/SearchContent.tsx
